+++ Humongous File Editor/src/functions/ChunkFunctions.cpp
@@ -12,9 +12,9 @@ namespace HumongousFileEditor
 	{
 		constexpr int LOOKUP_LIMIT = 15;
 
-		void ChunkFunctions::SetProgressBar(System::Windows::Forms::ToolStripProgressBar^ progressBar, size_t fvalue)
+		void ChunkFunctions::SetProgressBar(System::Windows::Forms::ToolStripProgressBar^ progressBar, double fvalue)
 		{
-			size_t value = floor(fvalue);
+			size_t value = static_cast<size_t>(floor(fvalue));
 			value = std::clamp(static_cast<int>(value), 0, 100);
 			size_t higherValue = value + 1;
 			higherValue = std::clamp(static_cast<int>(higherValue), 0, 100);
